feat: complete dashboard, discount, fix: order_detail (payment)

+++ src/main/java/com/qr_meal_web/service/impl/DiscountServiceImpl.java
@@ -1,5 +1,6 @@
 package com.qr_meal_web.service.impl;
 
+import com.qr_meal_web.enums.DiscountStatus;
 import com.qr_meal_web.model.Discount;
 import com.qr_meal_web.repository.DiscountRepository;
 import com.qr_meal_web.repository.impl.DiscountRepositoryImpl;
@@ -33,11 +34,22 @@ public class DiscountServiceImpl implements DiscountService {
 
     @Override
     public boolean delete(int id) {
+        boolean isExisted = new InvoiceServiceImpl().checkDiscountExisted(id);
+        if (isExisted) {
+            Discount discount = repo.getById(id);
+            discount.setStatus(DiscountStatus.fromCode(0));
+            return repo.update(discount);
+        }
         return repo.delete(id);
     }
 
     @Override
     public int getTotalQuantityDiscount() {
         return repo.getCountAllCustomer();
+    }
+
+    @Override
+    public List<Discount> getDiscountsLessThanPhone(String phone) {
+        return repo.getDiscountsLessThanPhone(phone);
     }
 }
